Move SensorUpdater.kt into sensors package.
