+++ app/src/main/java/io/homeassistant/companion/android/sensors/AllSensorsUpdater.kt
@@ -3,7 +3,6 @@ package io.homeassistant.companion.android.sensors
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import io.homeassistant.companion.android.SensorUpdater
 import io.homeassistant.companion.android.domain.integration.IntegrationUseCase
 
 abstract class AllSensorsUpdater(
