+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorUpdater.kt
@@ -1,4 +1,4 @@
-package io.homeassistant.companion.android
+package io.homeassistant.companion.android.sensors
 
 interface SensorUpdater {
     suspend fun updateSensors()
