+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorWorker.kt
@@ -8,7 +8,6 @@ import androidx.work.NetworkType
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
-import io.homeassistant.companion.android.SensorUpdater
 import io.homeassistant.companion.android.common.dagger.GraphComponentAccessor
 import io.homeassistant.companion.android.domain.integration.IntegrationUseCase
 import java.util.concurrent.TimeUnit
